requirements.txt

# AI Service Providers
# Core dependencies
# Data processing
# Docker integration
# GitHub integration
# PostgreSQL connection
# TEC_OFFICE_REPO requirements
# Testing and quality
# Utilities
# Utils
# Web interfaces and APIs
# WordPress
# WordPress and content management
anthropic>=0.2.0
anthropic>=0.6.0
beautifulsoup4>=4.11.0
colorama>=0.4.6
colorlog>=6.7.0
docker>=6.0.0
fastapi>=0.95.0
flake8>=6.0.0
flask>=2.0.0
google-generativeai>=0.3.0
gradio>=3.50.0
gunicorn>=21.2.0
huggingface_hub>=0.16.0
markdown>=3.4.0
numpy>=1.23.0
openai>=1.0.0
pandas>=1.5.0
psycopg2-binary
pydantic>=2.0.0
PyGithub>=1.58.0
pytest-cov>=4.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
python-wordpress-xmlrpc>=2.3
pyyaml>=6.0
requests>=2.28.0
tenacity>=8.2.0
tqdm>=4.65.0
tqdm>=4.66.0
typing>=3.7.4.3
uvicorn>=0.21.0
wordpress-api>=1.0
